Move Abstract Classes to Their own Package.
